requirements.txt
ultralytics
pytesseract
pillow
pdf2image
ollama
chromadb
numpy
tqdm